move code to KnoraUserToUserConverter
